Attempting to merge two different pull requests dealing with unicode.
I've also changed param paramteres to type short as that seems to be the better type. I don't get compiler complaints using longs or shorts, and others do for longs, so the best solution is just to use shorts all the time.

Also took a small but nice bind_parameter() simplification which seems to work fine.

Reverted the date format change.