Fix retrival of variable-length data in parts.
The termination condition of the loop in which SQLGetData
is called to retrieve varlen data in chunks was incomplete
and incorrect.
It only tested SQLGetData return code is greater than 0,
what led to infinite loop in case the call returned SQL_NO_DATA(100).

Test for this fix is not included, but it needs to be added soon.