test: Fix test_error with correct SQL State and native code
It seems the "Violation of PRIMARY KEY constraint" error is
accompanied/concluded with "The statement has been terminated"
as there is a hierarchy or list of diagnostics collected by
the `recent_error` utility function:
https://github.com/nanodbc/nanodbc/blob/946084a9a2a7a34429390c23c56bee2be4a2da6d/nanodbc/nanodbc.cpp#L391-L441

So, all the error messages are collected and concatenated into
a single message of `database_error` exception.
But, SQL State and SQL Native Code are stored for the last
diagnostics, that is, for "The statement has been terminated".

In future, this could be improved by changing `database_error`
to remember queue list of diagnostics.