test: Test unbinding vs access in order of increasing column number
This is to illustrate that unbinding only helps to work around some
cases of out-of-order access, namely, when  query interleaves
bound and unbound columns, but access is still performed
in order of increasing column number.

Unbinding does not avoid the limitation of SQLGetData
which requires access in order of increasing column number.

The limitation could be worked around with result adapter
caching values of all columns at once to ensure access in order
of increaseing column number.

Related to issue #228 and enhancement PR #236